Backend/requirements.txt
# BakeryOS Backend - Python Dependencies
# Created: March 22, 2026
# Django Version: 6.0.3
# Python Version: 3.10+
# Installation: pip install -r requirements.txt

# ============================================================================
# CORE FRAMEWORK & REST API
# ============================================================================
Django==6.0.3
djangorestframework==3.14.0
django-cors-headers==4.0.0

# ============================================================================
# AUTHENTICATION & SECURITY
# ============================================================================
djangorestframework-simplejwt==5.2.2
PyJWT==2.8.0
cryptography>=40.0.2

# ============================================================================
# DATABASE
# ============================================================================
# Use psycopg2-binary (pre-built) for development
# For production with PostgreSQL, use: psycopg2-binary==2.9.10 or newer
psycopg2-binary>=2.9.10

# ============================================================================
# API DOCUMENTATION & SCHEMA
# ============================================================================
drf-spectacular==0.26.1
PyYAML>=6.0.1

# ============================================================================
# FILTERING & SEARCHING
# ============================================================================
django-filter==23.1

# ============================================================================
# CONFIGURATION & UTILITIES
# ============================================================================
python-decouple==3.8
python-dotenv==1.0.0
Pillow>=9.5.0

# ============================================================================
# ASYNC & BACKGROUND TASKS
# ============================================================================
celery==5.3.1
redis>=4.5.5

# ============================================================================
# TESTING & QUALITY ASSURANCE
# ============================================================================
pytest==7.3.1
pytest-django==4.5.2
pytest-cov==4.1.0
factory-boy==3.2.1
faker==18.4.0

# ============================================================================
# DEVELOPMENT TOOLS
# ============================================================================
django-extensions==3.2.3
django-debug-toolbar==3.8.1
black==23.3.0
flake8==6.0.0
isort==5.12.0
pylint==2.17.4

# ============================================================================
# PRODUCTION DEPLOYMENT
# ============================================================================
gunicorn==20.1.0
whitenoise==6.4.0
dj-database-url==1.3.0

# ============================================================================
# TIMEZONE & DATE HANDLING
# ============================================================================
pytz==2023.3
python-dateutil==2.8.2

# ============================================================================
# DATA SERIALIZATION & EXPORT
# ============================================================================
openpyxl>=3.1.5
reportlab>=4.2.2

# ============================================================================
# OPTIONAL - UNCOMMENT IF NEEDED
# ============================================================================
# sentry-sdk==1.20.0                   # Error tracking
# django-anymail==9.0                  # Multi-provider email
# twilio==8.2.0                        # SMS notifications
# django-redis==5.2.0                  # Caching with Redis
# django-logging-json==1.16            # JSON logging format